Implements multithreaded benchmarking
At the moment, There appear to be some memory leaks, but I've
implemented benchmarking of the multithreaded version of introsort. I
also changed the type signature of the benchmark funtion so that I could
pass a partially initialized function pointer as a lambda, however, this
does mean that the benchmarking function isn't type safe (for function
pointers).